6.13： 修改新的 UI 前的备份。

+++ ManridyApp/BLETool/Model/SegmentedStepModel.m
@@ -12,7 +12,7 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"date = %@ AHCount = %ld CHCount = %ld stepNumber = %@ kCalNumber = %@ mileageNumber = %@ startTime = %@ timeInterval = %ld ", _date, _AHCount,  _CHCount, _stepNumber, _kCalNumber, _mileageNumber, _startTime, _timeInterval];
+    return [NSString stringWithFormat:@"date = %@ AHCount = %ld CHCount = %ld stepNumber = %@ kCalNumber = %@ mileageNumber = %@ startTime = %@ timeInterval = %ld ", _date, (long)_AHCount,  (long)_CHCount, _stepNumber, _kCalNumber, _mileageNumber, _startTime, (long)_timeInterval];
 }
 
 @end
